requirements.txt
# Python 3.12.11
numpy==2.0.2
pandas==2.2.2
scikit-learn==1.6.1
xgboost==3.0.4
torch==2.8.0+cu126
torch_geometric==2.6.1
matplotlib==3.10.0
seaborn==0.13.2
hdbscan==0.8.40
umap-learn==0.5.8